requirements.txt
python-telegram-bot==13.15
web3==6.10.0